Update Clikt to 5.0.2 and support both v1 and v2 version format
