+++ src/main/kotlin/dev/crmodders/versionsgen/Main.kt
@@ -1,6 +1,7 @@
 package dev.crmodders.versionsgen
 
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.main
 import com.github.ajalt.clikt.core.subcommands
 import dev.crmodders.versionsgen.command.FileCommand
